# Dokumentasi/Swagger/openAPI.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/webdeveloper.reski/portal/1.0.0
info:
  description: Portal with Golang and React
  version: "1.0.0"
  title: Portal
  contact:
    email: programmer.reski@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  /login:
    post:
      tags:
        - Auth
      summary: login user
      description: Login user bisa digunakan ketika user sudah registrasi
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestLogin'
            examples:
              userData1:
                description: "Data  user 1"
                value: {
                  username: "admin",
                  password: "admin123",
                }
              userData2:
                description: "Data  user 2"
                value: {
                  username: "reski",
                  password: "1234",
                }
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseLoginSuccess'
              examples:
                responseData:
                  description: response success to login to your account
                  value: {
                    code: 200,
                    message: "Login Success",
                    data: {
                      "user_id": 1,
                      "username": "admin",
                      "role": "admin",
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MzgzMjYxODAsInVzZXJJZCI6MSwidXNlcm5hbWUiOiJhZG1pbiJ9.AebFR-oQjUSOMez2ucDWkiMrS2eQIPmcYm5c71qZ_co"
                    }
                  }
                responseData2:
                  description: response success to login your account
                  value: {
                    code: 200,
                    message: "Login Success",
                    data: {
                      "user_id": 2,
                      "username": "reski",
                      "role": "creator",
                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MzgzMjYxODAsInVzZXJJZCI6MSwidXNlcm5hbWUiOiJhZG1pbiJ9.AebFR-oQjUSOMez2ucDWkiMrS2eQIPmcYm5c71qZ_co"
                    }
                  }
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData:
                  description: response failed to login to account
                  value: {
                    code: 400,
                    message: "email or password incorrect"
                  }
        '500':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData:
                  description: response failed to login to account
                  value: {
                    code: 500,
                    message: "Internal server error"
                  }
  /users:
    post:
      tags:
        - Users
      summary: Registration User
      description: Saat registrasi user harus input semua field
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/requestRegistrasi'
            examples:
              userData1:
                description: "Data  user 1"
                value: {
                  fullname: "thanos",
                  username: "madtitan",
                  email: "thanos@gmail.com",
                  password: "than321",
                  role: "admin",
                }
              userData2:
                description: "Data  user 2"
                value: {
                  fullname: "batman boy",
                  username: "batman",
                  email: "notbatman@gmail.com",
                  password: "wayne1234",
                  role: "creator",
                }
      responses:
        '200':
          description: Success Response
          content:
            application/json:
              examples:
                responseData:
                  description: response success to register user
                  value: {
                    code: 200,
                    message: "success to insert data",
                  }
                
        '400':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData:
                  description: response failed to register user
                  value: {
                    code: 400,
                    message: "all input data must be filled",
                    data : null
                  }
        '500':
          description: Failed Response
          content:
            application/json:
              examples:
                responseData:
                  description: response failed to register user
                  value: {
                    code: 500,
                    message: "Internal Server Error",
                    data : null
                  }
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  schemas:
    requestLogin:
      type: object
      properties:
        "username":
          type: string
          minLength: 1
          maxLength: 50
        "password":
          type: string
          minLength: 1
          maxLength: 50
        "role":
          type: string
          minLength: 1
          maxLength: 50
    responseLoginSuccess:
      type: object
      properties: {
        
        "message": {
          type: string
        },
        "data":{
          type: object
        }
      }
    requestRegistrasi:
      type: object
      properties:
        "username":
          type: string
          minLength: 1
          maxLength: 50
        "fullname":
          type: string
          minLength: 1
          maxLength: 50
        "password": 
          type: string
          minLength: 1
          maxLength: 50
        "email":
          type: string
          minLength: 1
          maxLength: 50
        "role":
          type: string
          minLength: 1
          maxLength: 15
...